% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{constructBasis}
\alias{constructBasis}
\title{constructBasis}
\usage{
constructBasis(
  splitMidPointY,
  y = NULL,
  splineDf = 2,
  basis,
  newKnots = F,
  eps = 0.001
)
}
\arguments{
\item{splitMidPointY}{vector of representative response values in each bin after discretization.}

\item{y}{vector of responses.}

\item{splineDf}{the number of response basis functions.}

\item{basis}{a character or a function specifying sufficient statistics, i.e., spline basis. For \code{basis = "Gaussian"}, y, y^2 are used. For \code{basis = "nsTransform"}, transformed natural cubic splines are used. If \code{basis} is a function, it should take a vector of response values and output a basis matrix: each row stands for a response value and each column stands for a basis function.}

\item{newKnots}{if \code{TRUE}, quantiles of \code{y} are used as the knots for splines. If \code{FALSE}, quantiles of \code{splitMidPointY} are used as the knots for splines.}

\item{eps}{a non-negative scalar added to the \code{penalty} for computational stability. Default is 0.001.}
}
\value{
This function returns a list of values.
\itemize{
\item z: the spline basis matrix.
\item penalty: a vector of penalty.factors.
}
}
\description{
This function construct response basis functions.
}
