% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{constructSplitPoint}
\alias{constructSplitPoint}
\title{constructSplitPoint}
\usage{
constructSplitPoint(X, numberSplit)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row represents an observation vector.}

\item{numberSplit}{split numbers for each covariate (length nvars). Each variable's range is divided into \code{numberSplit-1} intervals containing approximately the same number of observations.}
}
\value{
The function returns a candidate split list (length nvars). Each element is a vector corresponding to a certain variable's candidate splits (length of \code{numberSplit} or 1 + the number of unique values).
}
\description{
This function computes the list of candidate covariate splits.
}
