% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Lattice plot of conditional densities}
\usage{
densityPlot(
  X,
  minY = NULL,
  maxY = NULL,
  nGrid = 100,
  yGrid = NULL,
  model = NULL,
  trueDensity = NULL
)
}
\arguments{
\item{X}{the covariate matrix to predict conditional densities at; each row represents an observation vector.}

\item{minY}{the minimal possible response value. If \code{yGrid} is NULL, then the response grid will be generated by dividing \code{(minY, maxY)} into \code{nGrid} equal bins and using the mid-points as \code{yGrid}. Default is NULL.}

\item{maxY}{the maximal possible response value. For more, see \code{minY}. Default is NULL.}

\item{nGrid}{the number of response grid points. For more, see \code{minY}. Default is 100.}

\item{yGrid}{the response grid to evaluate conditional densities at. Default is NULL.}

\item{model}{a conditional density model with a \code{predict} method, e.g., a LinCDE boosting model. The call \code{predict(model, X, y)} should return a vector of estimated conditional densities f(yi | Xi). To only plot the true conditional densities, leave the \code{model} blank. Default is NULL.}

\item{trueDensity}{true conditional density function. The call \code{trueDensity(X, y)} should return a vector of true conditional densities f(yi | Xi). To only plot LinCDE's estimated conditional densities, leave the \code{trueDensity} blank. Default is NULL.}
}
\value{
The function outputs a lattice plot of conditional densities.
}
\description{
This function produces conditional density plots at input covariate values.
}
