% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dfToLambda}
\alias{dfToLambda}
\title{dfToLambda}
\usage{
dfToLambda(z, counts, splineDf, df, numberBin, penalty = NULL)
}
\arguments{
\item{z}{sufficient statistics matrix, of dimension \code{numberBin} x \code{splineDf}.}

\item{counts}{vector indicating the number of responses in each bin, of length \code{numberBin}.}

\item{splineDf}{the number of sufficient statistics.}

\item{df}{degrees of freedom.}

\item{numberBin}{the number of bins for response discretization.}

\item{penalty}{separate penalty factors applied to each coefficient, of length \code{splineDf}.}
}
\value{
The function returns the regularization parameter.
}
\description{
This function computes the regularization parameter corresponding to the given degrees of freedom.
}
