% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinCDE.boost.R
\name{LinCDE.boost}
\alias{LinCDE.boost}
\title{LinCDE.boost}
\usage{
LinCDE.boost(
  y,
  X,
  splitPoint = 20,
  basis = "nsTransform",
  splineDf = 10,
  minY = NULL,
  maxY = NULL,
  numberBin = 40,
  df = 4,
  penalty = NULL,
  prior = "Gaussian",
  depth = 1,
  n.trees = 100,
  shrinkage = 0.1,
  terminalSize = 20,
  alpha = 0.2,
  subsample = 1,
  centering = FALSE,
  centeringMethod = "randomForest",
  verbose = TRUE
)
}
\arguments{
\item{y}{response vector, of length nobs.}

\item{X}{input matrix, of dimension nobs x nvars; each row represents an observation vector.}

\item{splitPoint}{a list of candidate splits of length nvars or a scalar/vector of candidate split numbers. If \code{splitPoint} is a list, each object is a vector corresponding to a variable's candidate splits (including the left and right end points). The list's objects should be ordered the same as \code{X}'s columns. An alternative input is candidate split numbers, a scalar if all variables share the same number of candidate splits, a vector of length nvars if variables have different numbers of candidate splits. If candidate split numbers are given, each variable's range is divided into \code{splitPoint-1} intervals containing approximately the same number of observations. Default is 20. Note that if a variable has fewer unique values than the desired number of intervals, split intervals corresponding to unique values are created.}

\item{basis}{a character or a matrix specifying sufficient statistics, i.e., spline basis. For \code{basis = "Gaussian"}, y, y^2 are used. For \code{basis = "nsTransform"}, transformed natural cubic splines are used. Default is "nsTransform".}

\item{splineDf}{the number of sufficient statistics/spline basis. If \code{z = "Gaussian"}, \code{splineDf} is set to 2. Default is 10.}

\item{minY}{the user-provided left end of the response range. Default is NULL.}

\item{maxY}{the user-provided right end of the response range. Default is NULL.}

\item{numberBin}{the number of bins for response discretization. Default is 40. The response range is divided into \code{numberBin} equal-width bins.}

\item{df}{approximate degrees of freedom. \code{df} is used for determining the ridge regularization parameter. If \code{basis = "Gaussian"}, no penalization is implemented. If \code{df = splineDf}, there is no ridge penalization. Default is 6.}

\item{penalty}{vector of penalties applied to each sufficient statistics' coefficient.}

\item{prior}{a character or a function specifying initial carrier density. For \code{prior = "uniform"}, the uniform distribution over the response range is used. For \code{prior = "Gaussian"}, the Gaussian distribution with the marginal response mean and standard deviation is used. For \code{prior = "LindseyMarginal"}, the marginal response density estimated by Lindsey's method based on all responses is used. The argument \code{prior} can also be a homogeneous or heterogeneous conditional density function. The conditional density function should take a covariate matrix X, a response vector y, and output a vector of conditional densities f(yi | Xi). See the LinCDE vignette for examples. Default is "Gaussian".}

\item{depth}{the number of splits of each LinCDE tree. The number of terminal nodes is \code{depth + 1}. If \code{depth = 1}, an additive model is fitted. Default is 1.}

\item{n.trees}{the number of trees to fit. Default is 100.}

\item{shrinkage}{the shrinkage parameter applied to each tree in the expansion, value in (0,1]. Default is 0.1.}

\item{terminalSize}{the minimum number of observations in a terminal node. Default is 20.}

\item{alpha}{a hyperparameter in (0,1] to early stop the boosting. A smaller \code{alpha} is more likely to induce early stopping. If \code{alpha = 1}, no early stopping will be conducted.}

\item{subsample}{subsample ratio of the training samples in (0,1]. Default is 1.}

\item{centering}{a logical value. If \code{TRUE}, a conditional mean model is fitted first, and LinCDE boosting is applied to the residuals. The centering is recommended for responses whose conditional support varies wildly. See the LinCDE vignette for examples. Default is \code{FALSE}.}

\item{centeringMethod}{a character or a function specifying the conditional mean estimator. If \code{centeringMethod = "linearRegression"}, a regression model is fitted to the response. If \code{centeringMethod = "randomForest"}, a random forest model is fitted. If \code{centeringMethod} is a function, the call \code{centeringMethod(X)} should return a conditional mean model with a predict function. Default is "randomForest". Applies only to \code{centering = TRUE}.}

\item{verbose}{a logical value. If \code{TRUE}, progress and performance are printed. Default is \code{TRUE}.}
}
\value{
This function returns a LinCDE object consisting of a list of values.
\itemize{
\item trees: a list of LinCDE trees.
\item importanceScore: a named vector measuring the contribution of each covariate to the objective.
\item splitMidPointY: the vector of discretized bins' mid-points.
\item z: the spline basis matrix.
\item zTransformMatrix: the transformation matrix (of dimension splineDf x splineDf) multiplied by the standard natural cubic spline basis if \code{basis = "nsTransform"}.
\item prior: the prior function. The call \code{prior(X, Y)} should return a vector of prior conditional densities f(yi | Xi).
\item basis/depth/shrinkage/centering/centeringMethod: values inherited from the input arguments. If \code{centering} is \code{FALSE}, no \code{centeringMethod} is returned.
\item centeringModel: a centering model with a predict function. If \code{centering} is \code{FALSE}, no \code{centeringModel} is returned.
}
}
\description{
This function implements LinCDE boosting: a boosting algorithm of conditional density estimation with shallow LinCDE trees as base-learners.
}
