% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinCDEBoosting.R
\name{LinCDEBoostingHelper}
\alias{LinCDEBoostingHelper}
\title{LinCDEBoostingHelper}
\usage{
LinCDEBoostingHelper(
  X,
  yIndex,
  z = NULL,
  currDepth,
  tree,
  indexList,
  currNodeIndex,
  cellProb,
  improvementQueue,
  splitPoint,
  numberBin = 20,
  order = 4,
  lambda = 0,
  penalty = NULL,
  df = 2,
  alpha = 0.2,
  terminalSize = 20,
  depth = 2,
  shrinkage = 0.1
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row represents an observation vector.}

\item{yIndex}{discretized response vector, of length nobs.}

\item{z}{sufficient statistics matrix, of dimension \code{numberBin} x \code{order}. Default is NULL.}

\item{currDepth}{the current depth of the LinCDE tree.}

\item{tree}{the current LinCDE tree.}

\item{indexList}{the list of candidate splits.}

\item{currNodeIndex}{the index of the current node.}

\item{cellProb}{cell probability matrix, of dimension nobs x \code{numberBin}.}

\item{improvementQueue}{the queue recording improvements of candidate splits.}

\item{splitPoint}{a list of candidate splits, of length nvars.}

\item{numberBin}{the number of bins for response discretization. Default is 20.}

\item{order}{the number of sufficient statistics. Default is 4.}

\item{lambda}{the regularization parameter for Poisson regression. Default is 0.}

\item{penalty}{separate penalty factors applied to each coefficient, of length \code{order}. Default is NULL.}

\item{df}{degrees of freedom. Default is 2.}

\item{alpha}{parameter for the stopping rule of splitting. Default is 0.2.}

\item{terminalSize}{the minimum number of observations in a terminal node. Default is 20.}

\item{depth}{the number of splits of each LinCDE tree. The number of terminal nodes is \code{depth + 1}. For \code{depth = 1}, an additive model is fitted. Default is 2.}

\item{shrinkage}{the shrinkage parameter applied to each tree in the expansion, value in (0,1]. Default is 0.1.}
}
\value{
trees and cellProb.

This function returns a LinCDE tree \code{trees} and the updated cell probability matrix \code{cellProb}.
}
\description{
This function is a helper for the function LinCDEBoosting. This helper grows one LinCDE tree with heterogeneous carrier density.
}
