% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinCDECdf}
\alias{LinCDECdf}
\title{LinCDECdf}
\usage{
LinCDECdf(cellProb, y, splitPointY)
}
\arguments{
\item{cellProb}{matrix of cell probabilities, of dimension nobs x number of bins.}

\item{y}{values to compute cdf. at (length nobs).}

\item{splitPointY}{split points of responses (length number of bins + 1).}
}
\value{
This function return a vector of cdf. (length nobs).
}
\description{
This function computes the cdf. for a LinCDE boosting model in C++.
}
