% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{LinCDEExample}
\alias{LinCDEExample}
\title{Example of LinCDE boosting}
\usage{
LinCDEExample(setting = "GLM", m = 2)
}
\arguments{
\item{setting}{data generating mechanism type. For \code{setting = "GM"}, a locally Gaussian mixture conditional density model is used. For \code{setting = "GLM"}, a locally Gasussian conditional density model is used. More details can be found in the paper. Default is "GLM".}

\item{m}{the number of simulation trials. Default is 2.}
}
\value{
The function prints the relative improvement of the log-likelihood and displays conditional density plots in different subregions.
}
\description{
This function implements a toy example of LinCDE boosting.
}
