% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinCDEPredict.R
\name{LinCDEPredict}
\alias{LinCDEPredict}
\title{LinCDE prediction}
\usage{
LinCDEPredict(X, y = NULL, trees, splitPointYTest = NULL)
}
\arguments{
\item{X}{input matrix for prediction, of dimension nobs x nvars; each row represents an observation vector.}

\item{y}{response vector for prediction, of length nobs.}

\item{trees}{a LinCDE boosting model.}

\item{splitPointYTest}{vector of response break points. Default is NULL.}
}
\value{
The function returns a list of values.
\itemize{
\item density: conditional density predictions at \code{y}, vector of length nobs. If \code{y = NULL}, return \code{NA}.
\item testLogLikelihood: average test log-likelihood at \code{y}. If \code{y = NULL}, return \code{NA}.
\item cellProb: cell probability prediction matrix, of dimension nobs x number of discretization bins; each row represents a cell probability vector.
\item yDiscretized: vector of discretized responses/response residuals, of length number of discretization bins. If \code{splitPointYTest = NULL}, discretized responses/response residuals from the LinCDE boosting model \code{trees} are used.
}
}
\description{
This function makes predictions from a LinCDE boosting model.
}
