% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinCDEQuantiles}
\alias{LinCDEQuantiles}
\title{LinCDEQuantiles}
\usage{
LinCDEQuantiles(cellProb, probs, splitPointY)
}
\arguments{
\item{cellProb}{matrix of cell probabilities, of dimension nobs x number of bins.}

\item{probs}{levels of quantiles.}

\item{splitPointY}{split points of responses (length number of bins + 1).}
}
\value{
This function return a matrix of quantiles (dimension nobs x  number of quantile levels).
}
\description{
This function computes the quantiles for a LinCDE boosting model in C++.
}
