% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinCDESplit}
\alias{LinCDESplit}
\title{LinCDESplit}
\usage{
LinCDESplit(X, yIndex, cellProb, z, covMatrixInv, splitPoint, numberBin)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row represents an observation vector.}

\item{yIndex}{discretized response vector, of length nobs.}

\item{cellProb}{cell probability matrix, of dimension nobs x \code{numberBin}.}

\item{z}{sufficient statistics matrix, of dimension \code{numberBin} x number of sufficient statistics.}

\item{covMatrixInv}{inverse of psi'' (number of sufficient statistics x number of sufficient statistics).}

\item{splitPoint}{a candidate split list (length nvars). Each element is a vector corresponding to a certain variable's candidate splits (including the left and right end points).}

\item{numberBin}{the number of bins for response discretization.}
}
\value{
The function returns \code{splitVar}: the index of the variable to split at; \code{splitVal}: the cut-point of the split; \code{improvement}: the contribution of the split to the objective.
}
\description{
This function computes the approximately optimal split in C++. A split is invalid if the sample size on the left or right of the split falls below 10
}
