% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.LinCDE.R
\name{LinCDETreePredict}
\alias{LinCDETreePredict}
\title{LinCDETreePredict}
\usage{
LinCDETreePredict(X, tree, currentRow = 1)
}
\arguments{
\item{X}{input matrix for prediction, of dimension nobs x nvars; each row represents an observation vector.}

\item{tree}{a LinCDE tree.}

\item{currentRow}{an integer specifying the node position in \code{tree}. If the \code{currentRow}-th row of the \code{tree} is not a terminal node thus a split, the function \code{LinCDETreePredict} is applied to the left and right children of the split. If the \code{currentRow}-th row of the \code{tree} is a terminal node, predictions are computed.}
}
\value{
The function returns a list of values.
\itemize{
\item membership: length n vector: each element represents the row (terminal node) that a sample falls into.
}
}
\description{
This function conducts prediction of a LinCDE tree.
}
