% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LinCDE.R
\name{print.LinCDE}
\alias{print.LinCDE}
\title{print.LinCDE}
\usage{
\method{print}{LinCDE}(x, ...)
}
\arguments{
\item{x}{a LinCDE model (object).}

\item{...}{other parameters passed to print.default.}
}
\value{
The function prints a list of values.
\itemize{
\item model: a LinCDE model. The response is denoted by \code{y}. If the covariates' names are not provided in the training stage, the covariates will be named in the format "Xj".
\item hyperparameters including depth, shrinkage, basis, centering (centeringMethod if centering is \code{TRUE}), and function of the prior are provided.
\item importanceScore: relative influences of the covariates to the response distribution.
}
}
\description{
This function displays basic information about a LinCDE model (object).
}
