% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LinCDE.R
\name{summary.LinCDE}
\alias{summary.LinCDE}
\title{summary.LinCDE}
\usage{
\method{summary}{LinCDE}(
  object,
  cBars = length(object$importanceScore),
  plotit = TRUE,
  order = TRUE,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{object}{a LinCDE boosting model.}

\item{cBars}{the number of bars to plot. If \code{order=TRUE} the only the variables with the \code{cBars} largest relative influence will appear in the barplot. If \code{order=FALSE} then the first \code{cBars} variables will appear in the plot. In either case, the function will return the relative influence of all of the variables.}

\item{plotit}{an indicator as to whether the plot is generated.}

\item{order}{an indicator as to whether the plotted and/or returned relative influences are sorted.}

\item{normalize}{if TRUE then returns the normalized influence summing to \code{100}.}
}
\value{
This function returns a data frame where the first component is the variable name and the second is the computed relative influence/importance.
}
\description{
This function provides the relative covariate influence/importance of a LinCDE boosting model.
}
