% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Lattice plot of estimated densities}
\usage{
densityPlot(
  X,
  minY = 0,
  maxY = 0,
  trees = NULL,
  trueDensity = NULL,
  yMean = NULL
)
}
\arguments{
\item{X}{covariate matrix to predict conditional densities at; each row represents an observation vector.}

\item{minY}{the minimal possible response value. If \code{trees} is not NULL, then the minimal value of the \code{trees}'s response grid will be used. Default is 0.}

\item{maxY}{the maximal possible response value. If \code{trees} is not NULL, then the maximal value of the \code{trees}'s response grid will be used. Default is 0.}

\item{trees}{a LinCDE boosting model. To only plot the true conditional densities, leave the \code{trees} blank. Default is NULL.}

\item{trueDensity}{true conditional density function. The true conditional densities at a covariate response pair (X,y) can be computed as \code{trueDensity(X,y)}. To only plot LinCDE's estimated conditional densities, leave the \code{trueDensity} blank. Default is NULL.}

\item{yMean}{conditional mean vector. The estimated conditional densities will be shifted to center at \code{yMean}. Default is NULL (no shift).}
}
\description{
This function produces conditional density plots at input covariate values.
}
