% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{importanceScore}
\alias{importanceScore}
\title{importanceScore}
\usage{
importanceScore(model, d, n.trees = 1, var.names = NULL)
}
\arguments{
\item{d}{the number of covariates.}

\item{n.trees}{the number of trees in the LinCDE boosting model.}

\item{tree}{a LinCDE boosting model.}
}
\value{
The function returns a length \code{d} vector of covariate importances.
}
\description{
This function computes the importance score for a LinCDE boosting model.
}
