% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinCDEPredict.R
\name{LinCDETreePredict}
\alias{LinCDETreePredict}
\title{LinCDETreePredict}
\usage{
LinCDETreePredict(X, tree, currentRow = 1)
}
\arguments{
\item{X}{input matrix for prediction, of dimension nobs x nvars; each row represents an observation vector.}

\item{tree}{a LinCDE tree.}

\item{currentRow}{the row index of the current node.}
}
\value{
The function returns the \code{membership} vector (length nobs): each element represents the row (terminal node) that a sample falls into.
}
\description{
This function conducts prediction of a LinCDE tree.
}
