% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LinCDE.R
\name{summary.LinCDE}
\alias{summary.LinCDE}
\title{summary.LinCDE}
\usage{
\method{summary}{LinCDE}(
  object,
  ...,
  cBars = length(object$importanceScore),
  plotit = TRUE,
  order = TRUE,
  normalize = TRUE,
  main = NULL
)
}
\arguments{
\item{object}{a LinCDE model.}

\item{...}{other parameters.}

\item{cBars}{the number of bars to plot. If \code{order=TRUE} the only the variables with the \code{cBars} largest relative influence will appear in the barplot. If \code{order=FALSE} then the first \code{cBars} variables will appear in the plot. In either case, the function will return the relative influence of all of the variables.}

\item{plotit}{an indicator as to whether the plot is generated.}

\item{order}{an indicator as to whether the plotted and/or returned relative influences are sorted.}

\item{normalize}{if \code{TRUE} then returns the normalized influence summing to \code{100}.}

\item{main}{a main title for the plot.}
}
\value{
If the \code{object$centering} is \code{FALSE}, this function returns a data frame where the first component is the variable name and the second is the computed relative influence/importance. If the \code{object$centering} is \code{TRUE}, this function returns two data frames \code{location} and \code{beyondLocation}. In each data frame, the first component is the variable name and the second is the computed relative influence/importance. The \code{location} data frame records the relative influence/importance from \code{object$centeringModel}, and the \code{beyondLocation} data frame records the relative influence/importance from the LinCDE model fitted to the residuals of \code{object$centeringModel}.
}
\description{
This function provides relative covariate influences/importances of a LinCDE model.
}
